requirements.txt
ruff
rich
python-dotenv

langgraph
langgraph-api
langgraph-checkpoint
langgraph-sdk
langgraph-cli
langgraph-checkpoint-postgres
langchain-openai
langchain-community

firecrawl-py
playwright
# tavily-search

psycopg[binary,pool]
redis
celery
html-to-markdown
pandas
numpy
sqlmodel
